+++ app/routes/data-api/project/index.js
@@ -378,6 +378,22 @@ router.post('/:projectUrl/user/del', async function(req, res, next) {
     });
 });
 
+router.post('/:projectUrl/remove', function(req, res, next) {
+    res.type('json');
+
+    if(!req.haveAccess(req.project.project_id, "delete project")) {
+        next(new APIError('You do not have permission to delete this project'));
+        return;
+    }
+    model.projects.removeProject({
+        project_id: req.project.project_id,
+        external_id: req.body.external_id,
+        idToken: req.session.idToken
+    }).then(function() {
+        res.json({ result: 'success' });
+    }).catch(next);
+});
+
 router.get('/:projectUrl/user-permissions', function(req, res, next) {
     res.type('json');
     model.users.getPermissions(