+++ scripts/db/012-projects-table-add-deleted-column.sql
@@ -1,0 +1,2 @@
+ALTER TABLE `arbimon2`.`projects`
+ADD COLUMN `deleted_at` DATETIME DEFAULT NULL;